Alias mode to export a share to other namespaces
Although persistent volumes are cluster-wide, claims are namespaced. Sometimes it is desirable to expose the same NFS folder to an arbitrary number of namespaces. See two different issues in the old repo[1].

This PR introduces a new alias mode, in which the provisioner can expose to others an existing NFS claim in namespace X that it mounts. Then, the provisioner lets the administrator or the users create new claims for the same NFS tree in namespaces A, B, C, etc.

More detailed example:

1. The admin creates PVC1 in NS1 as before. If on AWS, they could be using efs-provisioner. Or, if it's NFS subdir external provisioner itself, a new directory will be created, just once. It shouldn't matter which one is used here, as long as it points to some NFS share.

1. They run nfs-client-provisioner in NS1 in alias mode (setting the `NFS_SERVER` variable to `--alias`), as a deployment named something like `nfs-alias-provisioner`, which mounts PVC1 and has its `PROVISIONER_NAME` variable set to `nfs-alias`.

1. They create a new StorageClass, `nfs-alias`, which uses the new `nfs-alias` provisioner.

1. One or more `nfs-alias` claims are created in namespaces A, B, C, etc. These could be created manually or as part of whatever automated namespace provisioning that the cluster administrator has setup.

1. Each of these claims will now use the same server and path as PVC1. The provisioner does that by looking up its own mount entry for PVC1 and passing the details through to the new PV.

So, in order to export the same folder to N namespaces, N+1 PVCs and two provisioners will be needed.

Some NFS provisioners might remove recursively the PVC1 directory tree if the claim gets deleted and retainPolicy == Delete|Recycle. Thus, aliases might see data disappear if storage is not set up differently.

This change tries as much as possible NOT to disrupt existing setups.

The code is not very complicated. The real work is mostly a matter of agreeing on how to trigger this mode, how to test it and what kind of guarantees to offer (or not to offer). Questions still left:

1. Is "alias" the right term? Previous choices I tossed out: clone (easy to mix with real volume cloning), proxy (there's no middle-man NFS server), passthrough (too nebulous).
1. Is NFS_SERVER=--alias the best way to trigger this? I like it because you can't set a real server name by mistake, you are forced into making a choice. And server names can't start with dashes.
1. Should one provisioner be able to handle both regular and alias volumes? If so, this could be done by registering a second provisioner named after the ALIAS_PROVISIONER_NAME variable. It would also need to look up or track the parent PVs for their `server:path`.
1. Should the new volume's reclaim policy be set to Retain?
1. Is it possible to prevent the original directory tree from getting delete as long as any alias are still around?
1. Should the auto-detection of the server host:path happen always, even when not in alias mode? It's a convenience feature that saves you from setting NFS_SERVER/NFS_PATH (at the cost of having to mount an existing volume).
1. What if we want a similar behaviour, where we inherit the details from PVC1, but each new claim gets its own subdirectory?
1. Should the parent's ReadOnly field be propagated to the newly provisioned PV?
1. Should this be made more generic so that one deployment can expose N different claims/trees to the other namespaces? It could work like this:

Say we have existing claims: `data`, `code`, `docs`. We want them available in any number of other namespaces, as allowed by RBAC. We run a foo.com/nfs-alias provisioner that mounts the three claims under /persistentvolumes/ and we use it in storage class `shared-volumes`.
Now we create claim `code` in namespace `developer1` under the new class. The provisioner will lookup the `server:path` for `/persistentvolumes/code` and stick those in a new volume `developer1-code-BLABLABLA`.

[1]
https://github.com/kubernetes-retired/external-storage/issues/1210
https://github.com/kubernetes-retired/external-storage/issues/1275